refactor: Dialog 패키지 경로 변경 및 구조 개선
