--- android/Bottari/presentation/src/main/java/com/bottari/presentation/common/dialog/DialogPresetType.kt
@@ -1,5 +1,0 @@
-package com.bottari.presentation.common.dialog
-
-enum class DialogPresetType {
-    EXIT_WITHOUT_SAVE,
-}
