+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/view/common/SafeArgument.kt
@@ -1,0 +1,6 @@
+package com.bottari.presentation.view.common
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+
+inline fun <T> Fragment.safeArgument(block: Bundle.() -> T): T? = runCatching { requireArguments().block() }.getOrNull()
