+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/view/common/alart/DialogListener.kt
@@ -1,4 +1,4 @@
-package com.bottari.presentation.common.dialog
+package com.bottari.presentation.view.common.alart
 
 interface DialogListener {
     fun onClickPositive()
